fix: move deep_update function to trtllm folder (#3070)
